@@ -4,23 +4,23 @@
   "private": true,
   "homepage": "https://batidrei.github.io/profile/",
   "dependencies": {
-    "@testing-library/jest-dom": "5.14.1",
-    "@testing-library/react": "12.0.0",
-    "@testing-library/user-event": "13.2.1",
-    "axios": ">=0.21.1",
-    "bootstrap": "5.0.2",
-    "gh-pages": "3.2.3",
-    "node-sass": "6.0.1",
-    "prop-types": "15.7.2",
-    "react": "17.0.2",
+    "@emotion/react": "^11.14.0",
+    "@testing-library/jest-dom": "6.9.1",
+    "@testing-library/react": "16.3.0",
+    "@testing-library/user-event": "14.6.1",
+    "axios": ">=1.12.2",
+    "bootstrap": "5.3.8",
+    "gh-pages": "6.3.0",
+    "prop-types": "15.8.1",
+    "react": "^18.2.0",
     "react-anchor-link-smooth-scroll": "1.0.12",
-    "react-bootstrap": "1.6.1",
-    "react-dom": "17.0.2",
-    "react-reveal": "1.2.2",
-    "react-router-dom": "5.2.0",
-    "react-scripts": "4.0.3",
+    "react-awesome-reveal": "^4.3.1",
+    "react-bootstrap": "2.10.10",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "7.9.4",
+    "react-scripts": "5.0.1",
     "readme-md-generator": "1.0.0",
-    "styled-components": "5.3.0"
+    "styled-components": "6.1.19"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -46,6 +46,7 @@
     ]
   },
   "devDependencies": {
-    "cypress": "8.0.0"
+    "cypress": "15.5.0",
+    "sass": "^1.93.2"
   }
 }
