@@ -1,20 +1,24 @@
-import React from 'react'
-import './scss/main.scss'
-import { HashRouter as Router, Route, Switch } from 'react-router-dom'
-import { Home } from './components/pages/home'
-import { Navigation } from './components/organims/navigation'
-import { error404 } from './components/pages/error404'
+// src/App.jsx
+import React from 'react';
+import './scss/main.scss';
+import { HashRouter as Router, Routes, Route } from 'react-router-dom';
+
+import { Home } from './components/pages/home';
+import { Navigation } from './components/organims/navigation';
+// Asegúrate de exportarlo en PascalCase desde el archivo
+// export const Error404 = () => ( ... )
+import { error404 } from './components/pages/error404';
 
 const App = () => (
   <Router>
-    <div className='back--in-down'>
+    <div className="back--in-down">
       <Navigation />
-      <Switch>
-        <Route path='/' exact component={Home} />
-        <Route component={error404}/>
-      </Switch>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="*" element={<error404 />} />
+      </Routes>
     </div>
   </Router>
-)
+);
 
-export default App;
+export default App;
